+++ src/prop.ts
@@ -20,7 +20,6 @@ import {
   DecoratedPropertyMetadata,
   DecoratedPropertyMetadataMap,
   MapPropOptions,
-  PropOptions,
   PropOptionsWithValidate,
   WhatIsIt
 } from './types';
@@ -262,15 +261,13 @@ export function _buildPropMetadata(input: DecoratedPropertyMetadata) {
 
         return;
       case WhatIsIt.MAP:
-        // "default" is a reserved keyword, thats why "_default" is used
-        const { default: _default }: PropOptions = options;
-        delete options.default;
-        delete options.of;
+        const mapped = utils.mapOptions(rawOptions, Type, target, key, true);
+
         schemas.get(name)[key] = {
           ...schemas.get(name)[key],
+          ...mapped.outer,
           type: Map,
-          default: _default,
-          of: { type: Type, ...options }
+          of: { type: Type, ...mapped.inner }
         };
 
         return;
@@ -313,13 +310,13 @@ export function _buildPropMetadata(input: DecoratedPropertyMetadata) {
 
       return;
     case WhatIsIt.MAP:
+      const mapped = utils.mapOptions(rawOptions, Type, target, key, false);
+
       schemas.get(name)[key] = {
         ...schemas.get(name)[key],
+        ...mapped.outer,
         type: Map,
-        of: {
-          type: virtualSchema,
-          ...options
-        }
+        of: { type: virtualSchema, ...mapped.inner }
       };
 
       return;
@@ -377,6 +374,7 @@ export function prop(options: PropOptionsWithValidate = {}) {
 export function mapProp(options: MapPropOptions) {
   return (target: any, key: string) => {
     const Type = options?.of;
+    delete options.of;
 
     if ('items' in options) {
       logger.warn('You might not want to use option "items" in a @mapProp, use @arrayProp (%s.%s)', utils.getName(target), key);